[refactor] Moved 'removeElementById' to its own TS file
