+++ src/SaveObject.js
@@ -19,12 +19,13 @@ import {loadSpecialServerIps, SpecialServerIps} from "./SpecialServerIps";
 import {loadStockMarket, StockMarket}           from "./StockMarket";
 import {dialogBoxCreate}                        from "../utils/DialogBox";
 import {gameOptionsBoxClose}                    from "../utils/GameOptions";
-import {clearEventListeners, removeElementById} from "../utils/HelperFunctions";
+import {clearEventListeners}                    from "../utils/HelperFunctions";
 import {Reviver, Generic_toJSON,
         Generic_fromJSON}                       from "../utils/JSONReviver";
 import {createElement}                          from "../utils/uiHelpers/createElement";
 import {createPopup}                            from "../utils/uiHelpers/createPopup";
 import {formatNumber}                           from "../utils/StringHelperFunctions";
+import {removeElementById}                      from "../utils/uiHelpers/removeElementById";
 
 import Decimal                                  from "decimal.js";
